requirements.txt
praw>=7.7.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.0.0
httpx>=0.27.0
pytest>=8.0.0
